Various reworkings.
Add a function to release/unclaim a job on the claimer object and
expose that functionality in the job object. Add an await() method
on the job which will do basic spinning (later functionality can
do more advanced types of waiting). Add a useful str() function to
the job so that it can be easily printed with some useful info about
it.