A few adjustments to the progress code
Remove an unneeded variable in the tasks class as
well as add a test for capturing and checking on
the progress that is emitted and adjust how the task
action should use a single function for doing state
changes (and another one for saving/resetting). Also
add in progress boundary checking (between 0 and 1).

Change-Id: Ic8cf56e3b25a493129f370b7b52e3cb2238da844