Remove weakref usage
If a connection escapes the scope of the
backend object, then python will actually
garbage collect the backend object even
though the connection object has a property
that allows you to fetch the backend. If the
property is used after the gc occurs a weakref
failure occurs.

Python is smart enough to correctly deallocate
these types of object links by itself so lets
just let it do that.

Fixes: bug 1228968

Change-Id: I49d36b74f896bd1be6c7a1a373b07bdaded3ef4b