Create database from models for SQLite
Alembic does not support SQLite, and we don't recommend to use SQLite
in production deployments, so migrations are rarely needed for SQLite.
So we don't bother about working around SQLite limitations, and create
database from models when it is in use. Of course, we still use
alembic to migrate schema on 'real' databases.

Change-Id: I8237a29c7bf5b60136bb1b72fd029fe9ce1abe1f