Avoid setting object variables
Instead of setting a object state variable, which makes
a graph_action object now have state we can instead just
have a new event boolean that will be set to true when
a suspension is detected. This value will then be used to
alter the status returned from the parallel graph action.

Also fixes a lack of returned status when there are no
items to actually run (aka the deps counter dictionary
is empty).

Change-Id: Ifa3ba7293613ba1e70e77d29504ea214f0b4c9e9