Fix entrypoints being updated/created by update.py
Previously entrypoints were not being created in the
target projects setup.cfg which means that the target
project would not be able to correctly use the taskflow
provided entrypoints. To fix this add a new helper
function to update.py that alters the entrypoint and
creates (or updates) the setup.cfg file where those
entrypoints are defined.

Fixes bug: 1235988

Change-Id: Ie32324c88e3c3642e58f3532e868f171f973d15e