Do not erase task progress details
Before this change we used to loose any progress details set by task
when it was updating its progress when task state was changed from
RUNNING to SUCCESS (and thus progress forced to 1.0). This commit
fixes this, so that when progress is updated from the engine latest
progress details are preserved.

Breaking change: to allow details to be associated with progress
value they were set for we save progress and then return progress
value with details.

Change-Id: Ic90e61ee3dcf62731696a0f10134bc448e7d6384