Ensure that mysql traditional mode is enabled
To avoid *silent* truncation of columns in mysql
it is much safer to turn on traditional mode which
behaves in the normal manner.

From the mysql docs:

Make MySQL behave like a "traditional" SQL database system.
A simple description of this mode is "give an error instead
of a warning" when inserting an incorrect value into a column.

Change-Id: Ic62af8514c377d2b8c934449116498855e03d7bd