Explicitly list the valid transitions to RESUMING state
Instead of having a catch-all in check_flow_transition
that handles valid transitions to the RESUMING state we
should just have an explicit mapping of which old_state to
RESUMING state is valid. This is much easier to examine and
reason about then having to determine which states the
catch-all will allow/be affected by.

Change-Id: I3a4226b17bf5ad8e5453a3ff6bc9078f7c24b520