Test using mysql + postgres if available
If mysql is avaiable via the standard testing user
'openstack_citest' with password 'openstack_citest'
then attempt to setup mysql or postgres and use
those databases for the duration of the test.

Adjust the tox template to include the two main mysql
library variants [MySQL-python, PyMySQL] and the psycopg2
postgres library so that these new tests will find the
appropriate library (sqlalchemy will attempt to find
these when connecting).

Change-Id: I90e59504663277d277d94e1f2a11031119bcedd6