Move some common/to be shared kazoo utils to kazoo_utils
Stopping and closing in a way that will avoid the current
kazoo bug will be a common utility function so move it to
kazoo_utils. The other common utility function will be version
checking so move that as well to kazoo_utils (and improve it
to allow for checking against a max version).

Change-Id: I81dfd9c01a00d7491b973d65d3b38f61b29dd298