Use listener instead of AutoSuspendTask in test_suspend_flow
Instead of injecting engine into storage (which is not clean, not
generally possible, and is an ugly hack) tests for suspension and
resumption now use SuspendingListener which suspends engine
when task with given name reaches given state. This gives us
cleaner and more portable across engines and persistence backends
solution for testing suspension.

Change-Id: Ieca80994bf4e2b6ee3a6249b644365136858bdd0