Add a delay before releasing the lock
To avoid when the lock gets released and the
time to acquire is below the float tolerance
add-on a sleep to ensure that there is always
some variation in timing before the acquisition
by another thread.

If the cause of this bug is a time shift (ntpd?)
then this will not be a fix (a fix for that is
only available in py3.3).

Fixes bug 1288521

Change-Id: I3294e81ce043a2c91730bb614020bb79c867dcdb