Rewrite assertion for same elements in sequences
New version is able to compare sequences whith elements that are not
hashable and cannot be compared (so that sorted() does not work).
For that, a utility function that caluclates difference between two
sequences was added. This function was also used in retry.ForEachBase.

The assertion was also renamed to assertItemsEqual, which matches
same assertion that was added in python 2.7.

Change-Id: I2b1b811190e9dc51718e4ca17ffc5c9015c34dc4