Adjust doc linking
Instead of linking to other topics (which then uses
the name of those topics as the link name) define a
name that makes more sense for the inline usage and
retain the link to the document using wording that
fits the surronding text.

Also adjust the futures/executor links to point to the
external links documenting these features.

Change-Id: I5a89e2f747dfec2505947f25c124b157271c07cf