Use a common message dispatcher
Instead of recreating a dispatcher in the server
and executor objects use a common dispatcher that
is shared between them. It will dispatch based on
the message type received into a provided dict of
dispatch handler callbacks.

It also can generically requeue messages and can
reject messages if they are missing key required
message properties ('type' in the current case).

Part of blueprint wbe-message-validation

Change-Id: I8320f4707183f36e6a69f0552cf62f99a5467b7e