Make version.py handle pbr not being installed
Since pbr is now only a dev/build-time requirement of taskflow
and not a run-time or test-time requirement we should not have
a version file that explicitly requires pbr to exist to function
correctly.

To accommodate when pbr is not found use the pkg_resources function
that can provide the installed version instead so that we still
provide back a valid version.

Change-Id: Id191d2b38def54b95a3467b4023a9540c284660d