Update the state graph builder to use state machine type
To build the state diagrams switch from using a custom
algorithm that builds a state machine in a table to using
our state machine class to build those same machines before
translation into a dot diagram.

This also allows us to use the state machine that already
is being used in the action engine as the source of transitions
to build the dot diagram with.

Change-Id: I4a0d16a3fb7c620c2774b535ab952b5d5006e9e9