Use oslotest to provide our base test case class
The oslotest library has a nice openstack testing integrated
base class that can ensure we setup our base test case using
the right logging fixtures, test timeouts, and output fixtures
that better operate in the openstack ecosystem.

This will also allow us to remove some of the functionality
that we currently have in our base test case and replace it with
the equivalent (or better) functionality that oslotest now
provides.

Part of blueprint use-oslo-test

Change-Id: I1602d5180ec8649a1899185972750ddddf65990f