Use the features that the oslotest mock base class provides
Instead of having our own mock subclass that has similar functions
as the oslotest base mocking class just use the base class functions
where we can and add on our own customizations as we choose to.

This change moves to using the base classes fixtures and also adjusts
the customized subclass method names to match closer to the rest of
the unittest classes method name style (camel-case not underscores).

Change-Id: If24530c0381d7fb99797acaa582d3be1d7054185