Allow loading conductors via entrypoints
Add an easy to use helper function that is exposed to
fetch a conductor in a way that matches how engines
and persistence backends and jobboards are loaded.

This also adjusts the single threaded conductor to
be now named blocking conductor and exposes an entrypoint
backend fetch() function to load current and future
conductors.

Change-Id: Id146d847c329d3e8510a8f24c3ec8b918680ddb5