Reduce the worker-engine joint testing time
This adjustment tweaks how workers and an engine use
there configuration (making it so they share the majority
of configuration). This also allows both engine and worker
to share the same configuration, which allows them to share
the same polling interval and so-on.

This has the effect of reducing the test time on a local
machine that I am running them on (one of many that I run
them on).

Before: Ran 1022 tests in ~130s (same machine)

After: Ran 1022 tests in ~60s (same machine)

Change-Id: I2150d26fb52cb86d61682124d710ba15c7ff27a6