Allow the worker banner to be written to an arbitrary location
Instead of always writing the worker startup banner to LOG.info
make it so that a callback can be provided that will receive the
worker banner and can write it to a desired location that the
callback specifies (or drop it completely). This is useful for those
who use workers but have an alternate location they desire the
banner to go (for example some arbitrary stream).

Change-Id: I61a4b7e9dd33ee06137caaed33153f5b3fbeb661