Rework pieces of the task callback capability
Unifies the bind, unbind, autobind parameters.

Also to make it easier to introspect what are a tasks associated
callbacks and events are provide a listeners_iter() method that
can be used to introspect the registered (event, callbacks) pairs that
are registered with a task.

Also adds more useful docstrings to the various callback associated
binding, unbinding functions to make it more understandable how they
are used and what they are provided.

Also makes the currently only default provided event 'update_progress'
a constant that can be referenced from the task module, which allows others
to easily find it and use it.

Change-Id: I14181a150b74fbd97f6ea976723f37c0ba4cec36