Add a jobboard high level architecture diagram
Incorporate a diagram into the documentation which
shows at a high level what the different components
of a jobboard are and the different actors that are
typically involved in using it.

Also moves the graffle (omnigraffle) xml/text files
to be tarballs; to save space and to avoid false line
counts in gerrit and stackalytics.

Change-Id: I5ddb0bdb7e2eb39844d064546a34795f08cf0c1f