Tweaks to setup.cfg
Add in 3.4 classifier (and adjust spacing of the
classifier section) since 3.4 support will be maintained
ongoing in the future.

Adjusts the keywords to be a smaller list of comma
separated values that summarize what taskflow is all
about.

Adds python2.6 as a minimum version so that it is well
understood (it was already implied) what the minimum
python version we currently support is.

Hopefully this adjustment fixes bug 1323731

Change-Id: Ia20120b3e4064fb315ddcf018588f3af4e503d1e