Use a module level constant to provide the DEFAULT_LISTEN_FOR
Instead of repeating the same tuple over and over again just
provide a module level constant in the listener base module that
defines the default listen_for tuple that is then used in all
subclasses as the default (if those subclasses support forwarding
the listen_for states in the first place).

This makes it so that users of listeners can more easily figure
out the default to pass, without having to get involved in the
inner details.

Change-Id: I3041db4d7eabfbe8d0b2cc267cfca0e46c0b7139