Move atom action handlers to there own subfolder/submodule
Since action handlers are atom specific it seems appropriate
to place the handlers under a specific submodule so that the
action engine module/folder layout is more obvious and makes
more sense.

This also adds on a handles(atom) method to those handlers
so that the runtime module/class can use those methods to
determine which handler should handle resetting an atom to
its initial state.

Change-Id: I7067d91347b41613fba1492d81d68a590f22c467