Remove usage of listener base postfix
The base prefix added on to the listener classes is
not very useful and is already understood that these are
base classes by documentation or abc.ABCMeta usage so
we don't need to specifically name these classes bases
to also show that they are a base class (useless duplicate
information).

So in this change we deprecate the 'ListenerBase' and the
'LoggingBase' and rename these classes to more appropriate
names.

Change-Id: Iaeaeaf698c23d71720ef7b62c8781996829e192a