Allow all deprecation helpers to take a stacklevel
To make the deprecation helpers more uniform in what
arguments/keyword arguments they take allow them all
to take a stacklevel keyword argument (which by default
is set to a good value that works).

This allows those users of this helper to override the
stacklevel in a uniform manner (if they so need to).

Change-Id: Icbfc115aa2aea4a5c01b99dcfa2a5c4e4785c458