Move the jobboard/job bases to a jobboard/base module
In order to match the directory/module layout of the
other pluggable backends better move the jobboard
modules that define the base abstract classes into
a single base file.

This makes it easier to look at the taskflow code-base
and understand the common layout.

This also makes the docs for the zookeeper jobboard
better and includes them in the generated developer docs
under a implementations section.

Change-Id: I36f29c37dcf2403782a75e45665bd7c0a146a06e