Avoid creating a temporary list(s) for tree type
Instead of creating a temporary list of the node using
its __iter__() function and then reversing that list just
use the natively provided reverse_iter() method instead
that reduces this wasteful list copying and creating in
the first place.

Also does the same in the pformat() function which was
needlessly creating a temporary list of children nodes
instead of just using the nodes __iter__() functionality
directly.

Change-Id: Ice4001e6d014d2c0a1f7d8b916c60370fd5443a7