Include the 'old_state' in all currently provided listeners
Since the key 'old_state' is now standard across all notifications
sent from flows, retries or tasks we can safely use it in the listeners
without worrying about hitting key errors when it does not exist/is not
provided.

This also adds an example which shows how to use the dynamic logging
listener to view this information.

Change-Id: If4456440674347a3fe4972d9d0fa16ba8373ef9f