Have the serial task executor shutdown/restart its executor
When the serial task executor is stopped, have it match what the
other executors do and shutdown its executor, and on start have
it restart/start it to be come functional again. This matches what
the other executors do and makes it easier to understand the common
pattern that is applied/used.

Change-Id: Id62b588b05262aa9e334a64f53e4c4a0d5b66159