Use a class provided logger before falling back to module
If a subclass overrides the logging listeners and provides
a class specific logger (under the attribute named _LOGGER)
try to use that before using the taskflow specific logging
module (and only use this one as a last resort).

This allows subclasses to easily override the default logger
without having to continually pass in a constructor argument
to do the same.

Change-Id: I91fea3db6cdd1dfb39963ff9589fd530fe087278