Use the enum library for the retry strategy enumerations
Instead of having a set of string constants that are the
retry strategy/decision result have it instead be an enumerator
that can be more clearly associated in docs and in code that
these values are explicitly the strategies that the retry
code can use (making it more obvious to users of these constants
what they are for).

This also updates the retry docs to use this new linkable
class when describing the various strategies that can be
used/returned.

Change-Id: I944e521562be26f5315d7553da8d5820fc284c49