Add + use read/write lock decorators
For when a whole function should be locked by
a read or write lock it is quite useful to have
decorators that acquire and release and wrap
the target function with the provided lock.

This change switches the storage module to use
these decorators where appropriate (mainly in
places where the read or write lock is activated
for the duration of the functions call, at which
point it is clearer to just use the new decorators
instead).

Change-Id: I70d2c1ab478e9b7da9446482a4ffb28f6f5227b7