Add support for conditional execution
To make it possible to alter the runtime flow via a simple
conditional like structure make it possible to have the graph
flow link function take a decider that is expected to be some
callable that will decide (via a boolean return) whether the
edge should actually be traversed when running. When a decider
returns false; the affected + successors will be set into the
IGNORE state and they will be exempt from future runtime and
scheduling decisions.

Part of blueprint taskflow-conditional-execution

Change-Id: Iab0ee46f86d6b8e747911174d54a7295b3fa404d