Always return scope walker instances from `fetch_scopes_for`
Instead of returning tuples with fully expanded scopes return walker
instances that internally know how to avoid recomputing the visible
scopes (they do this by caching each visibility level and looking in
the local cache before computing the scope and storing it in the cache).

This makes the usage more uniform and avoids returning different items
depending on what is found; making the code easier to follow and
understand.

Also makes the scope walker call to '_extract_atoms' to go via a static
method so that if it is ever desired to alter what '_extract_atoms' does
it can be more easily done (using standard inheritance).

Change-Id: I5916838163e6be843429fe7b89a0b5622e9c2f36