Make resolution/retry strategies more clear and better
This adds a separate method which is used to locate the
action and handler callback that will be used to resolve
the failure. This also adjusts the consultation of the
parent retry (if any) to ensure that we handle the no
parent retry case correctly.

Once a decision has been made; return it and add logging
that shows what is being activated and how many nodes were
affected by the resolution strategy (which can be useful to
know during debugging).

Change-Id: I28101765fce000dd7c56b7c3a1fbcf1a4315799b