Remove customized pyX.Y tox requirements
Instead of having custom pyX.Y that test a specific set of
requirements and similar with other tox envs, just enable
more test-requirements directly, in a later change we will
use environment markers to denote the differences.

Also tweaks the jobboard example to bypass/hack a needed eventlet
fix into it so that it actually runs without breaking/never finishing.
A TODO note has been left to eventually (someday?) remove this hack
when it is no longer needed.

Part of fix for bug 1445827

Change-Id: I0fb669f0bdfbe4fe81d89234173f2c3ac4d06de6