Add a new `ls_r` method
Instead of having a `ls` method that when used recursively
*always* returns the absolute path of items in the fake in
memory storage tree and *relative* paths (when used in
non-recursive mode) add a new `ls_r` method that can return
absolute *or* relative paths.

In the future it is highly likely that the the `ls` recursive
keyword argument will be removed (so preferring and
moving to the `ls_r` should occur earlier rather than
later), so this also adds a debtcollector removed keyword
argument decorator over the existing `ls` to ensure that users
are aware of this change (as well as a adjusted docstring).

Fixes bug 1458114

Change-Id: Id2a5869e94ac44679020a14297d1073d1dc2718f