Handle conductor ctrl-c more appropriately
When a conductor program is interrupted via ctrl-c
or equivalent it is much nicer log that that has happened
and to reraise that exception.

This also slightly tweaks the 99 bottles song to make it
even better, by having more pieces/tasks, which makes it
possible to kill the program during each task and see how
the resumption works when a flow is composed of segments.

Change-Id: I5d242eba9a043ef96646ba74ea5928daa0691ed0