Add smarter/better/faster impl. of `ensure_atoms`
Instead of ensuring an atom at a time we can examine
at bulk what atoms details need to be created, and then
only do any persistence operations if any atom details
are determined to not exist. This saves on unneeded
reads and writes and helps improve storage and prepare
times.

Part of blueprint make-things-speedy

Change-Id: I2e7e3cc60cc97cbbe4e66c69e22c34ee31ebfeb3